+++ hitchhiking_diary_server/components/images/endpoints/opengraph.py
@@ -2,7 +2,7 @@ import io
 from uuid import UUID
 
 import staticmaps
-from PIL import Image, ImageDraw, ImageFont, ImageEnhance, ImageFilter
+from PIL import Image, ImageDraw, ImageFont, ImageEnhance
 from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
 from starlette import status