+++ hitchhiking_diary_server/components/images/router.py
@@ -1,6 +1,8 @@
 from fastapi import APIRouter
 
 from .endpoints import opengraph
+from .endpoints import story
 
 router = APIRouter()
 router.include_router(opengraph.router)
+router.include_router(story.router)