+++ CHANGELOG.md
@@ -9,4 +9,5 @@ Initial release
 
 - **Added**: Sync API
 - **Added**: Simple explore portal
-- **Added**: OpenGraph Image generator
+- **Added**: OpenGraph Image Generator
+- **Added**: Instagram Story Generator